refactor: MemberRole 열거형 분리
