+++ src/main/java/roomescape/controller/MemberController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.api;
+package roomescape.controller;
 
 import java.util.List;
