+++ src/main/java/roomescape/domain/Member.java
@@ -5,17 +5,21 @@ public record Member(
         String name,
         String email,
         String password,
-        String role
+        MemberRole role
 ) {
 
-    private static final String DEFAULT_ROLE = "normal";
+    private static final MemberRole DEFAULT_ROLE = MemberRole.normal;
 
     public Member(String name, String email, String password) {
         this(null, name, email, password, DEFAULT_ROLE);
     }
 
     public Member(String name, String email, String password, String role) {
-        this(null, name, email, password, role);
+        this(null, name, email, password, MemberRole.valueOf(role));
+    }
+
+    public Member(Long id, String name, String email, String password, String role) {
+        this(id, name, email, password, MemberRole.valueOf(role));
     }
 
     public Member withId(Long id) {
