+++ src/main/java/roomescape/domain/MemberRole.java
@@ -1,0 +1,7 @@
+package roomescape.domain;
+
+public enum MemberRole {
+    normal,
+    admin,
+    ;
+}
